+++ nepali/number/__init__.py
@@ -1,5 +1,5 @@
 from ._number import NepaliNumber
-from .nepalinumber import nepalinumber
+from ._nepalinumber import nepalinumber
 from .utils import (
     add_comma,
     add_comma_english,
@@ -10,7 +10,6 @@ from .utils import (
 
 __all__ = [
     "NepaliNumber",
-    "nepalinumber",
     "add_comma",
     "add_comma_english",
     "convert_and_add_comma",
